day 10/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-dotenv==1.1.1
requests==2.32.4
python-multipart==0.0.9
google-generativeai==0.2.0
assamblyai==0.16.0
murf==0.1.0